Operating airlines are not marketing airlines
If it's a partner airline flight, i.e. you booked through United but are
flying Lufthansa, then the marketing airline will be United and the
operating airline Lufthansa. If it's not a partner flight then operating
airline will be empty and only marketing airline will be populated.

This PR rationalizes that behavior and selects the actual airline and
flight number in operation.

I also added the airline code and flight number to the title of the
calendar entry... I like this but your OCD may protest. Can change it if
you like.