Operating airlines are not marketing airlines (#4)
* Fixed ProfileEmailAddress type - it's not an array

* Changed references from TripIt token to password

This is a breaking change - you'll need to update your docker run et al
commands.

* Keep the past in the past

Added the `-past` flag. If specified, it'll return all past trips as
well as all future trips. It defaults to `false` and hence the bot will
only return future trips.

* Indentation - how rude

* Operating airlines are not marketing airlines

If it's a partner airline flight, i.e. you booked through United but are
flying Lufthansa, then the marketing airline will be United and the
operating airline Lufthansa. If it's not a partner flight then operating
airline will be empty and only marketing airline will be populated.

This PR rationalizes that behavior and selects the actual airline and
flight number in operation.

I also added the airline code and flight number to the title of the
calendar entry... I like this but your OCD may protest. Can change it if
you like.

* Updated .gitignore